robot-server/requirements.win.txt
fastapi
uvicorn[standard]
python-dotenv
opencv-python